Move compile-time related components to a separate compile-time page
